Added rotation and turn speeds
The rotation of the rover and of the wheels are now seperately defined
variables. The rotation speed is now variable and capped at a max speed.
Ackerman steering can no longer have it's wheels rotating 360 degrees as
this would be impossible in real life. It's now capped at a max of 60
degrees in each rotation.