Added crab rotation basis
I've created the base key input and structure to allow for crab steering
to rotate about its center. This will utilize the A/S keys on the
keyboard currently. I've started working on moving the wheels and rover
to correspond with this but at the moment it is buggy so it has been
omitted.