Fixed crab steering bug
Crab steering no longer has the problem of the wheels rotating when
pressing up or down to accelerate in the forward or reverse direction.
This means that pressing up will not rotate the wheels right causing the
rover to spin in a tight circle when it should cause it to go in a
straight line in the direction the wheels are currently pointing.