Rover rotates in crab; needs wheels. Code cleanup
The rover can now rotate in crab steering mode by pressing A and S.
However, this is not reflected in its current movement as I'm still
trying to get the wheels to work and didn't want to lose functionality
while working on that part. The code has been cleaned up removing unused
code and fixing indentation where needed.